Add Experimental Support for Zarr Version 3
I am putting up this PR just to get it on the record.  It is
still failing the github action workflows for mingw and for
Visual Studio for cmake. Despite my best efforts I cannot fix
them.

## Change Overview
* Add experimental support for the Zarr Version 3 Specification.
  This is a bit of a moving target, so it is subject to change.
  This required major re-factoring of the NCZarr code.
* As discussed in a recent netcdf meeting, convert NCZarr V2 to
  store all netcdf-4 specific info as attributes.  Backward
  read-compatibility with the older NCZARR format is maintained.
* Add a new directory -- v3_nczarr_tests -- to test out the Zarr V3
  implementation. This requires duplicating the tests in directory
  "nczarr_test" into the directory "v3_nczarr_tests".
  This is a bit of a complicated hack and is documented in docs/internal.md.
* Fix the -Wconversion and -Wsign-compare warnings in
  libnczarr, nczarr_test, and v3_nczarr_test.

## Specific Changes

### Automake and CMake Build Changes
* Convert the ENABLE_XXX options to NETCDF_ENABLE_XXX options
  (part of the cmake overhaul).
* Disable the ENABLE_NETCDF4/ENABLE_NETCDF_4 options in favor of ENABLE_HDF5.
  This will cause complaints.
* Remove the ENABLE_CLIENTSIZE_FILTERS options since has been disabled for a while.
* Add option to specify the default Zarr format to use when creating datasets.
  It is currently set to use Zarr Version 2.

### NCZarr Specific Changes
* Add an "hdf5raw" codec format for cases where an hdf5 filter is known
  but no corresponding codec format is defined.
* The Zarr/Codec API has been modified to support including
  a property list argument to pass information about which Zarr
  format (Version2 vs Version 2) is required.

### Misc. Unrelated Changes
* The remotetest server is currently unstable, so provide a mechanism
  for force disabling calls to remotetest.unidata.ucar.edu.
  This is enabled by adding a repository variable named
  REMOTETESTDOWN with the value "yes".
* Fix CMakeLists.txt to use the uname command as an alternate
  to using the hostname command (which does not work under cygwin).

## Notes
* The current refactor to handle Version 3 is a stop-gap.
  At some point, it needs fixing to remove as much common code
  as possible to minimize the size of zformat2.c and zformat3.c
* Additional refactoring will be needed to support the new array->array
  codecs.
* The NCZarr related CMakeLists.txt need work to conform to the new
  cmake build methods.

+++ libnczarr/zodom.h
@@ -9,7 +9,7 @@
 struct NCZSlice;
 
 typedef struct NCZOdometer {
-    int rank; /*rank */
+    size_t rank; /*rank */
     size64_t* start;
     size64_t* stride;
     size64_t* stop; /* start + (count*stride) */
@@ -24,8 +24,8 @@ typedef struct NCZOdometer {
 
 /**************************************************/
 /* From zodom.c */
-extern NCZOdometer* nczodom_new(int rank, const size64_t*, const size64_t*, const size64_t*, const size64_t*);
-extern NCZOdometer* nczodom_fromslices(int rank, const struct NCZSlice* slices);
+extern NCZOdometer* nczodom_new(size_t rank, const size64_t*, const size64_t*, const size64_t*, const size64_t*);
+extern NCZOdometer* nczodom_fromslices(size_t rank, const struct NCZSlice* slices);
 extern int nczodom_more(const NCZOdometer*);
 extern void nczodom_next(NCZOdometer*);
 extern size64_t* nczodom_indices(const NCZOdometer*);
